refactor: update args
